requirements.txt
pypubsub
requests
matplotlib
networkx
PyPDF2
PyMuPDF
PyYAML
pdftotext
python-dotenv
pytest
sphinx
sphinx-rtd-theme